SOF-1860 -migration runner container, helpers to rebuild and execute, schema versions scanner
