+++ database/manager/tsconfig.json
@@ -1,7 +1,7 @@
 {
-  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "sourceMap": false,
-    "noEmit": true
+    "noEmit": true,
+    "noEmitOnError": true,
   }
 }
