+++ database/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "runner",
+  "version": "1.0.0",
+  "private": true,
+  "scripts": {
+  },
+  "packageManager": "pnpm@8.6.6",
+  "dependencies": {
+  },
+  "devDependencies": {
+    "@types/node": "^20.5.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
+  }
+}
